Added first layer/round of consolidated nic backend
