+++ monarch_rdma/src/lib.rs
@@ -21,8 +21,10 @@ mod macros;
 pub mod backend;
 pub mod config;
 pub mod device_selection;
-pub mod efa;
-pub mod broadcom;
+///pub mod efa;
+///pub mod broadcom;
+pub mod nic_backend;
+pub mod vendors;
 pub mod local_memory;
 mod rdma_components;
 mod rdma_manager_actor;
